requirements.txt
Pillow
customtkinter
packaging
ImageHash
numpy
pyinstaller
tkinterdnd2
